Pull rewards information from database instead of ipfs
